requirements.txt
PySide6>=6.4.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0


# 打包时使用

Pillow

# PyInstaller在处理.ico文件时，需要用Pillow来验证和转换图标格式。
# 即使你的logo.ico已经是正确的ICO格式，PyInstaller仍会尝试加载验证它。